refactor: move ToolbarExtensions.kt to widgets
