+++ app/src/main/java/com/f0x1d/logfox/ui/fragment/ExtendedCopyFragment.kt
@@ -6,7 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.fragment.navArgs
 import com.f0x1d.logfox.databinding.FragmentExtendedCopyBinding
-import com.f0x1d.logfox.extensions.views.setupBackButtonForNavController
+import com.f0x1d.logfox.extensions.views.widgets.setupBackButtonForNavController
 import com.f0x1d.logfox.ui.fragment.base.BaseFragment
 import dev.chrisbanes.insetter.applyInsetter
