+++ app/src/main/java/com/f0x1d/logfox/ui/fragment/settings/base/BasePreferenceFragment.kt
@@ -5,7 +5,7 @@ import android.view.View
 import androidx.preference.PreferenceFragmentCompat
 import com.f0x1d.logfox.R
 import com.f0x1d.logfox.extensions.isHorizontalOrientation
-import com.f0x1d.logfox.extensions.views.setupBackButtonForNavController
+import com.f0x1d.logfox.extensions.views.widgets.setupBackButtonForNavController
 import com.google.android.material.appbar.MaterialToolbar
 import dev.chrisbanes.insetter.applyInsetter
